refactor proposal execution routes
